common/requirements.txt
python-decouple
pymongo